Move `species-egg-tiers.ts`
