+++ src/data/species-egg-tiers.ts
